Create logger with either deafult or custom logging config
I changed the intent of the next-logger.config.js to be a config instead of an instance,
therefore I can return the
logger instance using whichever config I used.

The other way to do this would be to use the instance the consumer passes, or initialise for them.

For now I'll createLogger despite the risk of the
consumer having an extra logger they don't need